refactor: extract cloneRepo into shared session-utils
Move the clone strategy (gh → SSH → HTTPS fallback) from start.ts into
a shared cloneRepo function in session-utils.ts. Both start.ts and
spawn.ts now use the same function, eliminating duplication flagged by
Bugbot.
