+++ packages/cli/src/commands/start.ts
@@ -29,7 +29,8 @@ import {
   type ProjectConfig,
   type ParsedRepoUrl,
 } from "@composio/ao-core";
-import { exec, execSilent } from "../lib/shell.js";
+import { exec } from "../lib/shell.js";
+import { cloneRepo } from "../lib/session-utils.js";
 import { getSessionManager } from "../lib/create-session-manager.js";
 import { ensureLifecycleWorker, stopLifecycleWorker } from "../lib/lifecycle-service.js";
 import {
@@ -111,46 +112,6 @@ function resolveProjectByRepo(
 
   // No repo match — fall back to standard resolution (works for single-project)
   return resolveProject(config);
-}
-
-/**
- * Clone a repo with authentication support.
- *
- * Strategy:
- *   1. Try `gh repo clone owner/repo target -- --depth 1` — handles GitHub auth
- *      for private repos via the user's `gh auth` token.
- *   2. Fall back to `git clone --depth 1` with SSH URL — works for users with
- *      SSH keys configured (common for private repos without gh).
- *   3. Final fallback to `git clone --depth 1` with HTTPS URL — works for
- *      public repos without any auth setup.
- */
-async function cloneRepo(parsed: ParsedRepoUrl, targetDir: string, cwd: string): Promise<void> {
-  // 1. Try gh repo clone (handles GitHub auth automatically)
-  if (parsed.host === "github.com") {
-    const ghAvailable = (await execSilent("gh", ["auth", "status"])) !== null;
-    if (ghAvailable) {
-      try {
-        await exec("gh", ["repo", "clone", parsed.ownerRepo, targetDir, "--", "--depth", "1"], {
-          cwd,
-        });
-        return;
-      } catch {
-        // gh clone failed — fall through to git clone with SSH
-      }
-    }
-  }
-
-  // 2. Try git clone with SSH URL (works with SSH keys for private repos)
-  const sshUrl = `git@${parsed.host}:${parsed.ownerRepo}.git`;
-  try {
-    await exec("git", ["clone", "--depth", "1", sshUrl, targetDir], { cwd });
-    return;
-  } catch {
-    // SSH failed — fall through to HTTPS
-  }
-
-  // 3. Final fallback: HTTPS (works for public repos)
-  await exec("git", ["clone", "--depth", "1", parsed.cloneUrl, targetDir], { cwd });
 }
 
 /**
